requirements.txt
beautifulsoup4==4.4.1
python-dateutil==2.4.2
requests==2.9.1
unicodecsv==0.14.1
